split recognizecard between express call and logic
