+++ server/package.json
@@ -4,7 +4,7 @@
 	"description": "",
 	"main": "server.js",
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
+		"test": "jest",
 		"start": "node server.js",
 		"dev": "nodemon server.js"
 	},
@@ -17,7 +17,8 @@
 		"express": "^4.21.1",
 		"multer": "^1.4.5-lts.1",
 		"node-fetch": "^2.7.0",
-		"pokemontcgsdk": "^2.0.0"
+		"pokemontcgsdk": "^2.0.0",
+		"jest": "^28.0.0"
 	},
 	"devDependencies": {
 		"nodemon": "^3.1.7"
